requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pillow==9.2.0
torch==1.12.1
torchvision==0.13.1
onnxruntime
onnxruntime-gpu
python-multipart
ultralytics
flask
ultralytics
Pillow
requests
torch==1.12.1
torchvision==0.13.1
ultralytics
imageio[avif]


